pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-google-vertexai>=3.2.0",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.6",
    "langchain-qdrant>=1.1.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
]
